+ latitude & longitude in SiteInspectionService latitude & longitude : long -> double

+++ CrowdServiceCore/src/edu/fudan/se/crowdservice/core/ConcreteService.java
@@ -23,7 +23,7 @@ import java.util.logging.Level;
  */
 public abstract class ConcreteService implements BundleActivator {
     protected Context context;
-    protected long longitude, latitude;
+    protected double longitude, latitude;
     String consumerId;
     int time, cost, actualCost;
     String templateName;
